Allowing only creator to EDIT or DELETE posts

+++ Workspace/mean-course/backend/routes/posts.routes.js
@@ -71,12 +71,18 @@ router.put(
     }
 
     Post.updateOne(
-      { _id: req.params.id },
+      { _id: req.params.id, creator: req.userData.userId },
       { title: req.body.title, content: req.body.content, imagePath: imagePath }
-    ).then(() => {
-      res.status(200).json({
-        message: "Post updated successfully.",
-      });
+    ).then((result) => {
+      if (result.nModified > 0) {
+        res.status(200).json({
+          message: "Post updated successfully.",
+        });
+      } else {
+        res.status(401).json({
+          message: "Unauthorized",
+        });
+      }
     });
   }
 );
@@ -123,11 +129,19 @@ router.get("", (req, res, next) => {
 // DELETE a post
 router.delete("/:id", checkAuth, (req, res, next) => {
   // fetch all posts from DB
-  Post.deleteOne({ _id: req.params.id }).then(() => {
-    res.status(200).json({
-      message: "Posts Deleted",
-    });
-  });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
+    (result) => {
+      if (result.n > 0) {
+        res.status(200).json({
+          message: "Posts Deleted",
+        });
+      } else {
+        res.status(401).json({
+          message: "Unauthorized",
+        });
+      }
+    }
+  );
 });
 
 module.exports = router;
